refactor: move + rename ModelTypeTag -> FeatureType
